backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Cache
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector Database
qdrant-client==1.6.9
sentence-transformers==2.7.0

# LLM Integrations
openai==1.3.7

# Data Processing
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiohttp==3.9.0

# Crypto/Bitcoin APIs
requests==2.31.0

# Text Processing & NLP
nltk==3.8.1

# Security & Auth
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1